backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
replicate==0.22.0
requests==2.31.0
mediapipe==0.10.9
opencv-python-headless==4.9.0.80
trimesh==4.0.10
pymeshlab==2025.7
numpy==1.26.3
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
